+++ tests/common.py
@@ -8,8 +8,8 @@ from jinja2 import Template
 import tempfile
 
 TESTS_HOME = os.path.abspath(os.path.dirname(__file__))
-PYWPS_CFG = os.path.join(TESTS_HOME, 'pywps.cfg')
-ROOCS_CFG = os.path.join(tempfile.gettempdir(), 'roocs.ini')
+PYWPS_CFG = os.path.join(TESTS_HOME, "pywps.cfg")
+ROOCS_CFG = os.path.join(tempfile.gettempdir(), "roocs.ini")
 
 VERSION = "1.0.0"
 WPS, OWS = get_ElementMakerForVersion(VERSION)
@@ -37,18 +37,17 @@ def write_roocs_cfg():
     base_dir = {{ base_dir }}/mini-esgf-data/test_data/group_workspaces/jasmin2/cp4cds1/vol1/data
     """
     cfg = Template(cfg_templ).render(base_dir=TESTS_HOME)
-    with open(ROOCS_CFG, 'w') as fp:
+    with open(ROOCS_CFG, "w") as fp:
         fp.write(cfg)
     # point to roocs cfg in environment
-    os.environ['ROOCS_CONFIG'] = ROOCS_CFG
+    os.environ["ROOCS_CONFIG"] = ROOCS_CFG
 
 
 def resource_file(filepath):
-    return os.path.join(TESTS_HOME, 'testdata', filepath)
+    return os.path.join(TESTS_HOME, "testdata", filepath)
 
 
 class WpsTestClient(WpsClient):
-
     def get(self, *args, **kwargs):
         query = "?"
         for key, value in kwargs.items():
@@ -64,19 +63,20 @@ def get_output(doc):
     """Copied from pywps/tests/test_execute.py.
     TODO: make this helper method public in pywps."""
     output = {}
-    for output_el in xpath_ns(doc, '/wps:ExecuteResponse'
-                                   '/wps:ProcessOutputs/wps:Output'):
-        [identifier_el] = xpath_ns(output_el, './ows:Identifier')
+    for output_el in xpath_ns(
+        doc, "/wps:ExecuteResponse" "/wps:ProcessOutputs/wps:Output"
+    ):
+        [identifier_el] = xpath_ns(output_el, "./ows:Identifier")
 
-        lit_el = xpath_ns(output_el, './wps:Data/wps:LiteralData')
+        lit_el = xpath_ns(output_el, "./wps:Data/wps:LiteralData")
         if lit_el != []:
             output[identifier_el.text] = lit_el[0].text
 
-        ref_el = xpath_ns(output_el, './wps:Reference')
+        ref_el = xpath_ns(output_el, "./wps:Reference")
         if ref_el != []:
-            output[identifier_el.text] = ref_el[0].attrib['href']
+            output[identifier_el.text] = ref_el[0].attrib["href"]
 
-        data_el = xpath_ns(output_el, './wps:Data/wps:ComplexData')
+        data_el = xpath_ns(output_el, "./wps:Data/wps:ComplexData")
         if data_el != []:
             output[identifier_el.text] = data_el[0].text