+++ CHANGES.rst
@@ -3,7 +3,9 @@ Changes
 
 Unreleased
 ==========
-* ``apply_fixes`` option added for WPS processes and the ``Operator`` class.
+* ``apply_fixes`` and ``original_files`` option added for WPS processes and the ``Operator`` class.
+* ``director`` module added. This makes decisions on what is returned - NetCDF files or original file URLs.
+* ``python-dateutil>=2.8.1`` added as a new dependency.
 
 0.2.0 (2020-11-19)
 ==================